refactor: Modify Folder Structure
