+++ src/App.styles.js
@@ -1,0 +1,27 @@
+import styled from 'styled-components';
+
+export const AppLayout = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  max-width: 1920px;
+`;
+
+export const HomeSection = styled.section`
+  padding-top: 5rem;
+`;
+
+export const HomeFolderName = styled.h2`
+  padding: 1rem 0;
+`;
+
+export const MainLayout = styled.main`
+  width: 100vw;
+  margin - top: 2rem;
+  background: #ffffff;
+`;
+
+export const MainBox = styled.div`
+  padding: 2rem 5rem;
+`;
