+++ src/components/SearchBar/SearchBar.jsx
@@ -1,0 +1,11 @@
+import * as S from './SearchBar.styles';
+
+const SearchBar = () => {
+  return (
+    <S.SearchBarSection>
+      <S.SearchBarInput placeholder='링크를 검색해 보세요.' />
+    </S.SearchBarSection>
+  );
+};
+
+export default SearchBar;
