--- src/components/UI/Serchbar.jsx
@@ -1,11 +1,0 @@
-import * as S from '../../styles/Searchbar.styles';
-
-const SearchBar = () => {
-  return (
-    <S.SearchbarSection>
-      <S.SearchbarInput placeholder='링크를 검색해 보세요.' />
-    </S.SearchbarSection>
-  );
-};
-
-export default SearchBar;
