--- src/styles/DefaultMain.js
@@ -1,20 +1,0 @@
-import styled from 'styled-components';
-
-export const DefaultMainSection = styled.section`
-  display: flex;
-  flex-direction: column;
-  gap: 3rem;
-  align-items: center;
-  justify-content: flex-end;
-  width: 100%;
-  max-width: 1920px;
-  padding: 5rem 6rem 0;
-  overflow: hidden;
-`;
-
-export const DefaultMainHomeText = styled.section`
-  font-weight: 700;
-  font-size: 3.5rem;
-  line-height: 4rem;
-  text-align: center;
-`;
